refactor: extract affine quotient
